create login and register temp

+++ base/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render,redirect
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.db.models import Q
@@ -45,8 +46,12 @@ def createroom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateroom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!=room.host:
+        return HttpResponse("You are not allowed to be here")
+
     print(room) 
     if request.method=="POST":
         form=RoomForm(request.POST,instance=room)
@@ -58,9 +63,13 @@ def updateroom(request,pk):
     return render(request,'base/room_form.html',context)
 
 
-    
+@login_required(login_url='login')
 def deleteroom(request,pk):
     room=Room.objects.get(id=pk)
+
+    if request.user!=room.host:
+        return HttpResponse("You are not allowed to be here")
+
     if request.method=="POST":
         room.delete()
         return redirect('home')
@@ -84,9 +93,24 @@ def loginpage(request):
         except:
             messages.error(request, "User not exist")
             print("Except")
-        
-    return render(request,'base/login_register.html')
+    return render(request,'base/login.html')
 
 def logoutpage(request):
     logout(request)
     return redirect('home')
+
+def register(request):
+    if request.method=="POST":
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            print(user)
+            return redirect('home')
+        else :
+            messages.error(request, "There is a problem")
+    form=UserCreationForm()
+    context={'form':form}
+    return render(request,'base/register.html',context)
